* adds BasicAuthValidator template * adds currentUser stub * moves all server files to dedicated server folder
