+++ test/example_test.dart
@@ -4,7 +4,6 @@ Future main() async {
   final harness = Harness()..install();
 
   test("GET /example returns 200 {'key': 'value'}", () async {
-    expectResponse(await harness.agent.get("/example"), 200,
-        body: {"key": "value"});
+    expectResponse(await harness.agent.get("/ping"), 200, body: 'pong');
   });
 }
